WIP: Oh what a mess of mixing a lexer with a parser
I knew it was stupid, yet I've to go solve the remaining issues.
Curly braces status across sequential lines is a nightmare.

Does a closing curly count if it is on a line which starts with
(or has at some position) and "end-of-line" comment, i.e. an
apostrophe, count as closing curly brace?

In the good news part of this commit here is that a found a
stupid mistake: never use a non-const reference to a type
if you don't intend to modify it.

Or otherwise an assignment will not set a new referece, but
__write to the reference__, as this is what you wrote.

In short alway write:
const YourType ref& = some_class->function(params);
and only write
YourType ref& som_class->function(params);
if you really intend to possibly replace the reference with
a new value.

This was what was causing the parsed array of tokens/strings
(QVector<P2Words>) to be modified where I did not intend it
to be modified.

Now for some more buf fixing elsewhere.

+++ mainwindow.h
@@ -87,7 +87,7 @@ private slots:
     void setAsmOpcodesHex();
     void incAsmFontSize();
     void decAsmFontSize();
-    void setAsmFontSize(qreal size);
+    void setAsmFontSize(int size);
 
     void setDasmOpcodes(int mode);
     void setDasmOpcodesBin();
@@ -96,7 +96,7 @@ private slots:
     void setDasmOpcodesHex();
     void incDasmFontSize();
     void decDasmFontSize();
-    void setDasmFontSize(qreal size);
+    void setDasmFontSize(int size);
 
     void setDasmLowercase(bool check);
     void asmHeaderColums(const QPoint& pos);
